added mentee and mentor links on buttons in new landing page, added fullblack color, replaced new landing page with index.tsx
