+++ src/components/pages-about/TeamMember/index.tsx
@@ -9,14 +9,14 @@ function index({ imgSrc = null, name, pronouns, role }) {
     <div className={styles.container}>
       <Avatar size={216} src={imgSrc} />
       <div className={styles.nameContainer}>
-        <Title level={4} className="textAlignCenter">
+        <Title level={4} className="textAlignCenter characterPrimaryBlack">
           {name}
         </Title>
         <Text className="footnote textNeutral8 textAlignCenter">
           ({pronouns})
         </Text>
         <div className={styles.roleContainer}>
-          <Text className="regularWeight">{role}</Text>
+          <Text className="regularWeight characterPrimaryBlack">{role}</Text>
         </div>
       </div>
     </div>
