+++ src/styles/colors.css
@@ -38,6 +38,9 @@
   --character-primary-inverse: #fff;
   --character-secondary: rgba(0, 0, 0, 0.55);
   --character-disabled-placeholder: rgba(0, 0, 0, 0.25);
+
+  /* Text Colors added in phase 4 */
+  --character-primary-black: rgba(0, 0, 0);
   --character-geek-blue7: #1d39c4;
 
   --border-radius-base: 12px;
